correção de endpoint e service

+++ Java/ecommerce-backend/src/main/java/com/alejandro/ecommerce/carrinho/CarrinhoService.java
@@ -5,44 +5,65 @@ import com.alejandro.ecommerce.pedido.PedidoService;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class CarrinhoService {
 
     private final CarrinhoRepository carrinhoRepository;
+    private final PedidoService pedidoService;
 
-    public CarrinhoService(CarrinhoRepository carrinhoRepository) {
+    public CarrinhoService(CarrinhoRepository carrinhoRepository,
+                           PedidoService pedidoService) {
         this.carrinhoRepository = carrinhoRepository;
+        this.pedidoService = pedidoService;
     }
 
-    public List<Carrinho> listarTodos() {
-        return carrinhoRepository.findAll();
+    public List<Carrinho> listarPorUsuario(Long usuarioId) {
+        return carrinhoRepository.findByUsuarioId(usuarioId);
     }
 
-    public Carrinho alterar(Carrinho carrinho) {
-        // validações podem entrar aqui
-        return carrinhoRepository.save(carrinho);
-    }
+    public Carrinho adicionar(Carrinho req, Long usuarioId) {
 
-    public Carrinho finalizar(Long usuarioId, PedidoService pedidoService) {
-        List<Carrinho> itens = carrinhoRepository.findByUsuarioId(usuarioId);
+        // 🔥 evita duplicação (muito importante)
+        Optional<Carrinho> existente =
+                carrinhoRepository.findByUsuarioIdAndProdutoId(
+                        usuarioId, req.getProduto().getId()
+                );
 
-        // cria o pedido com valor total
-        Pedido pedido = pedidoService.criarPedido(usuarioId, itens);
+        if (existente.isPresent()) {
+            Carrinho item = existente.get();
+            item.setQuantidade(item.getQuantidade() + req.getQuantidade());
+            return carrinhoRepository.save(item);
+        }
 
-        // limpar carrinho
-        carrinhoRepository.deleteAll(itens);
+            req.getUsuario().setId(usuarioId);
 
-        return pedido;
+        return carrinhoRepository.save(req);
     }
 
     @Transactional
-    public Carrinho remover(Long id, Long usuarioId) {
+    public void remover(Long id, Long usuarioId) {
         int deletados = carrinhoRepository.deleteByIdAndUsuarioId(id, usuarioId);
+
         if (deletados == 0) {
             throw new RuntimeException("Item não encontrado ou não pertence ao usuário");
         }
+    }
 
-        carrinhoRepository.deleteById(id);
+    @Transactional
+    public Pedido checkout(Long usuarioId) {
+
+        List<Carrinho> itens = carrinhoRepository.findByUsuarioId(usuarioId);
+
+        if (itens.isEmpty()) {
+            throw new RuntimeException("Carrinho vazio");
+        }
+
+        Pedido pedido = pedidoService.criarPedido(usuarioId, itens);
+
+        carrinhoRepository.deleteAll(itens);
+
+        return pedido;
     }
 }
